compartment_and_add_extra

add compartment label to the updated transcript data.frame returned by fastReseg_perFOV_full_process and optional to combine it with transcripts got trimmed and originally extracellular transcripts

Usage

compartment_and_add_extra(
  updated_transDF,
  transcript_df,
  updatedCellID_coln = "updated_cellID",
  compartment_coln = "transComp",
  combine_extra = FALSE
)

Arguments

updated_transDF

transcript data.frame with transcript id and updated cell segmentation, one of the outputs of fastReseg_perFOV_full_process

transcript_df

list of 2 elements, intraC and extraC for intracellular and extracellular transcript data.frame, respectively; columns contain transcript id and original cell segmentation, output of prepare_perFOV_transDF

updatedCellID_coln

column name of updated cell ID in updated_transDF, default = "updated_cellID"

compartment_coln

column name for transcript compartment label, default = "transComp"

combine_extra

flag to combine original extracellular transcripts and trimmed transcripts back to the updated transcript data.frame, slow process if many transcript in each FOV file. (default = FALSE)

Value

a transcript data.frame containing all original transcripts with updated cell segmentation and transcript compartments for intraC, extraC and trimmed