createSpatialDelaunayNW_from_spatLocs

generate delaunay network based on provided config and spatial location using GiottoClass functions

Usage

createSpatialDelaunayNW_from_spatLocs(
  config_spatNW = list(name = "spatial_network", dimensions = "all", method = "Delaunay",
    minimum_k = 0, delaunay_method = "delaunayn_geometry", maximum_distance_delaunay =
    "auto", options = "Pp", Y = TRUE, j = TRUE, S = 0),
  spatLocs_df,
  ID_column = "cell_ID",
  spatLocs_column = c("sdimx", "sdimy", "sdimz")
)

Arguments

config_spatNW

configuration list

spatLocs_df

data.frame for spatial location of each entry for cell or transcript

ID_column

column name for entry ID in spatLocs_df

spatLocs_column

column name for 1st, 2nd, optional 3rd dimension of spatial coordinates in spatLocs_df

Value

delaunay_network_Obj, a spatial network object created by GiottoClass functions

Details

This function leverages GiottoClass package to create spatial networks from spatial coordinates. An example config_spatNW list is shown below with possible options on controlling the spatial network generation. For more details, see the manual for GiottoClass::createSpatialNetwork. #' '

name

spatial network name; default = 'spatial_network'

dimensions

a vector for which spatial dimensions to use, default = 'all' to use all dimentions

method

method name for creating a spatial network, default = 'Delaunay'

minimum_k

minimum number of nearest neighbors if maximum_distance != NULL

delaunay_method

Delaunay method to use, choose from c("delaunayn_geometry", "deldir", "RTriangle"), default = "delaunayn_geometry"

maximum_distance_delaunay

distance cuttoff for nearest neighbors to consider for Delaunay network, default = "auto"

options

(geometry) String containing extra control options for the underlying Qhull command; see the Qhull documentation (../doc/qhull/html/qdelaun.html) for the available options; default = Pp, do not report precision problems)

Y

(RTriangle) If TRUE prohibits the insertion of Steiner points on the mesh boundary

j

(RTriangle) If TRUE jettisons vertices that are not part of the final triangulation from the output.

S

(RTriangle) Specifies the maximum number of added Steiner points.