igraph_delete_edges

Delete edges from an igraph object with version compatibility

Usage

igraph_delete_edges(graph, edges)

Arguments

graph

An igraph object from which edges will be deleted.

edges

A vector of edge IDs or an edge selector to delete.

Value

An igraph object with the specified edges removed.

Details

This function wraps igraph::delete_edges() and ensures compatibility with older versions (<2.0.0) of igraph that used delete.edges().