initializeAllSegRes

initialize result holder for fastReseg_full_pipeline based on preprocessed data

Usage

initializeAllSegRes(
  prep_res,
  save_intermediates = TRUE,
  return_perCellData = TRUE
)

Arguments

prep_res

nested list of elements for 'refProfiles', 'baselineData', 'cutoffs_list', outputs of runPreprocess

save_intermediates

flag to save intermediate outputs into output folder, including data.frame for spatial modeling statistics of each cell,

return_perCellData

flag to return and save to output folder for gene x cell count matrix and per cell DF with updated mean spatial coordinates and new cell type

Value

a list

refProfiles

a genes X clusters matrix of cluster-specific reference profiles used in resegmenation pipeline

baselineData

a list of two matrice in cluster X percentile format for the cluster-specific percentile distribution of per cell value; span_score is for the average per molecule transcript tLLR score of each cell, span_transNum is for the transcript number of each cell.

cutoffs_list

a list of cutoffs used in resegmentation pipeline, including, score_baseline, lowerCutoff_transNum, higherCutoff_transNum, cellular_distance_cutoff, molecular_distance_cutoff

ctrl_genes

a vector of control genes whose transcript scores are set to fixed value for all cell types

updated_perCellDT

an empty list, return when return_perCellData = TRUE

updated_perCellExprs

an empty list, return when return_perCellData = TRUE

reseg_actions

a list of 4 empty elements, same output format as decide_ReSegment_Operations, return when save_intermediates = TRUE