Get number of cores for parallelized operations

Get number of cores for parallelized operations

Usage

numCores(percentCores = 0.9, minNotUsedCores = 2)

Arguments

percentCores

percent of cores to use for parallelization, value range 0 to 1

minNotUsedCores

minimum number of cores to leave for background processes

Value

number of cores to use for mclapply