prepResegDF

supporting function for fastReseg_perFOV_full_process, combine runTranscriptErrorDetection() output with transcript data.frame to prep for runSegRefinement()

Usage

prepResegDF(
  transcript_df,
  groupDF_ToFlagTrans,
  cellID_coln = "UMI_cellID",
  transID_coln = "UMI_transID"
)

Arguments

transcript_df

the data.frame for each transcript with columns for transcript_id, target or gene name, original cell_id, spatial coordinates and tLLR_maxCellType.

groupDF_ToFlagTrans

data frame for transcripts in chosen_cells only, with columns for connect_group,tmp_cellID,group_maxCellType, output of runTranscriptErrorDetection()

cellID_coln

the column name of cell_ID in transcript_df

transID_coln

the column name of transcript_ID in transcript_df

Value

a list of two elements

reseg_transcript_df

data.frame with transcript_id, target/geneName, x, y, cell_id for all transcript groups tmp_cellID and the cell type of maximum transcript scores for each transcript group group_maxCellType

groups_to_reseg

vector of chosen transcript groups need to be evaluate for re-segmentation