transDF_to_perCell_data

get per cell expression matrix and optional metadata like mean spatial coordinates and cell types from transcript data.frame

Usage

transDF_to_perCell_data(
  transcript_df,
  transGene_coln = "target",
  cellID_coln = "updated_cellID",
  spatLocs_colns = c("x", "y", "z"),
  celltype_coln = "updated_celltype",
  return_cellMeta = FALSE
)

Arguments

transcript_df

transcript data.frame

transGene_coln

column name of target or gene name in transcript_df

cellID_coln

column name of cell_ID in transcript_df

spatLocs_colns

column names for 1st, 2nd and optional 3rd dimension of spatial coordinates in transcript_df

celltype_coln

column name for cell type in transcript_df

return_cellMeta

flag to return per cell data.table for metadata like mean spatial coordinates and cell types

Value

a list

perCell_DT

a per cell data.table with mean spatial coordinates, cell type, return when return_cellMeta = TRUE

perCell_expression

a gene x cell count sparse matrix derived from transcript data.frame after resegmentation