per cell gene expression matrix based on raw transcript data files generated by SMI

The example dataset is based on the cropped regions from two different FOVs of same SMI RNA readout run for FFPE melanoma tissue sample. The corresponding raw transcript data files are

extdata/Run4104_FOV001__complete_code_cell_target_call_coord.csv

extdata/Run4104_FOV002__complete_code_cell_target_call_coord.csv

Each raw transcript data files have 1 transcript per row with 25 different meta information in columns. Some important columns are

x

spatial coordinate of the transcript in x-axis of the given FOV, unit in pixel, 0.18um per pixel

y

spatial coordinate of the transcript in y-axis of the given FOV, unit in pixel, 0.18um per pixel

z

spatial coordinate of the transcript in z-axis of the given FOV, unit in z-step, 0.8um per z-step

target

gene identity of given transcript

CellId

the cell label assignment of given transcript within given FOV based on original cell segmentation

The per cell gene expression matrix is calculated from the raw transcript data files to include only the 960 true target genes without including control probes.

Usage

data(example_CellGeneExpr)

Format

An object of class "matrix" with 754 cells in row and 960 genes in columns

See also